[lifetime_annotations] Treat function type aliases consistently with other aliases.
This fixes a crash repro recently added to rs_bindings_from_cc.

The crash ultimately resulted from `Importer::ConvertType()` trying to descend
into the defintion of the type alias while the `ValueLifetimes` in the
`lifetimes` parameter did not contain lifetimes for the definition of the type
alias. As defense in depth, add an assertion to `Importer::ConvertType()` that
would have caught this inconsistency at an earlier stage.

PiperOrigin-RevId: 546825633
Change-Id: Ife2d4f12dfe41bae635d7c1516604cfb49d5df19