Refactor ReturnValueSlot to be a class instead of a union
Unions with ctors, dtors, and member functions are unusual. When readers see these features being used they are likely to assume they are looking at a class.

ReturnValueSlot was a union to allow manual control of the `value_` member. It is possible to achieve the same effect using an unnamed union, putting the `union` keyword next to the definition of `value_`, hopefully making the code more readable.

PiperOrigin-RevId: 546896718
Change-Id: Iac38200441cb8f28307b77d9c35ea42e744fd9ca