Replace is_known property with from_nullable
The old property meant "was the source either Nullable or Nonnull?"
The new property means "was the source Nullable?"

Despite how central these properties are, I believe this is a very safe change.

These behave in very similar ways. (As descriptions of the source, they don't
affect runtime behavior and are mostly just moved around by the framework).
The "is nullable" test is the same as before: in scenarios where the pointer is
null, a nonnull source is impossible, so known == from_nullable.

There are a couple of advantages motivating the change:

- from_nullable is IMO a more concrete concept and so easier to reason about
- inference uses variables for e.g. annotated_nullable and annotated_nonnull.
  mapping these to from_nullable is simpler, and yields a simpler SAT system.

PiperOrigin-RevId: 548656500
Change-Id: Ib1d3deb150ef55a7b4ebc15bb1ef91d6b5198c1f