Only keep `is_trivially_foo` asserts.
This CL stops generating the following `static_assert` flavours:

* `is_move_constructible`
* `is_move_assignable`
* `is_copy_constructible`,
* `is_copy_assignable`
* `is_destructible`.

This is okay, because these properties are still directly exposed through
the API inside the generated file (e.g. the destructor is declared as not `=delete`d).
This is desirable, because it simplifies some follow-up refactoring of
`format_adt`.

This CL still keeps the `static_assert` flavours that aren't part of
the API inside the generated file:

* `is_trivially_copy_constructible`
* `is_trivially_copy_assignable`
* `is_trivially_move_constructible`
* `is_trivially_move_assignable`
* `is_trivially_destructible`.

These `static_assert`s are an important part of an API contract and can't be
inferred from other API elements.  For example, a `=default`ed move constructor
doesn't necessarily imply guarantee trivial movability for arbitrary C++ types
because the "trivial" aspect depends on other factors.

PiperOrigin-RevId: 549051530
Change-Id: Icb9ab7b51b4030a1cf4a35f6343cffa28be8d383