Refactor `ReturnValueSlot.Test` to avoid reading destroyed objects.
The test wants to check that the object pointed by `slot_ptr` has been
destroyed.  Before this CL the test would read the `slot_ptr->state`
field and assert that it is equal to `kDestroyedState`.  Reading an
already destroyed object is UB and was caught by MSAN.  After this CL,
the assertion is done through `destroyed_locations` instead.

PiperOrigin-RevId: 551219313
Change-Id: I4a5e085ad6a7ae8943721b93e52197fd11245a1c