[nullability] Simplify flow conditions produced when comparing against `nullptr`.
When comparing against `nullptr`, instead of adding three implications to the
flow condition, we can leave the flow condition unchanged and simply use the
nullstate of the operand that isn't `nullptr`, potentially with an added
negation.

Comparing against `nullptr` is the overwhelmingly common case, and simpler flow
conditions help in two ways:

- The SAT solver needs to do less work.

- Flow conditions are easier to interpret when debugging.

The tests in test/comparisons.cc already provide coverage for the new case
distinctions introduced here.

PiperOrigin-RevId: 559659135
Change-Id: I37ad7430654646cd31c3bcd4e69e97010496d6e5