[nullability] Fix `ParenTypeInTemplate` test.
Before, it was using `targetNullable()` and `targetNonnull()` as function names,
but we only analyze functions called `target()`.

Framework change [#66197](https://github.com/llvm/llvm-project/pull/66197) fixes
this so that we error out if there is no function called `target()`.

This patch renames the functions to `target()` and additionally splits the test
apart so there is only one call to `checkDiagnostics()` in each test (which
makes it easier to understand what is going on if a test fails).

Unfortunately, it turns out that we've regressed on two lines in
`ParenTypeInTemplate3` while the test was not testing anything.

PiperOrigin-RevId: 566268964
Change-Id: Ifac39e5c157f378740430bd18bbb02f0e51f0cb5