Allow multiple headers in `--bindings-from-dependency`.
A rust crate will only ever have one header. However, if we export types that come from the Rust bindings for a C++ library, that C++ library may have multiple headers. And so if we want to support these round-tripped-types, using `--bindings-from-dependency` to record which headers they came from, `--bindings-from-dependency` must have a vec of includes, not just a single include.

Similarly, the provider which outputs this must support a list of headers, not just a single header -- but that is a separate CL. (And at any rate, we can't usefully _send_ multiple headers until an updated crubit release.)

This isn't a sufficient change to cc_bindings_from_rs to support round-tripping the interop (we still need to generate correct include paths and type names), but it's part of that change.

PiperOrigin-RevId: 570132807
Change-Id: I9fe02bc63cb35e46af46d3047ce5ac14a33e4e8b