Prepare `CcBindingsFromRustInfo` for round-tripped interop.
Specifically, allow for multiple headers per crate, and move it to a separate bazel file which we can use from rs_bindings_from_cc.

This doesn't otherwise do anything -- the binary itself will not accept multiple headers until unknown commit is released (it'll just pick one arbitrarily), and we do not actually _populate_ it with multiple headers (and won't until release), but this change is relatively standalone and helps keeps the changelists small.

PiperOrigin-RevId: 570161088
Change-Id: I7c250cc9d3b7b7a32d9097d5db2518e43073b224