[nullability] Fix crash in `transferFlowSensitiveAccessorCall()`.
Need to use `dyn_cast_or_null` instead of `dyn_cast` because `getValue()` may
return null. (This patch adds a test that triggers this case.)

The new test exposed another issue: In the case where the line discussed above
sets `PointerVal` to null, we need to execute the fallback that calls
`getPointerValueFromExpr()` to ensure that we initialize the `PointerValue`'s
nullability properties. (The diagnosis expects every `PointerValue` to have
nullability properties set.)

PiperOrigin-RevId: 576862372
Change-Id: Ib6f4e7811cac534cb41c5f969f73ef802993499b