Require that both sides of initialization/assignment are of supported types.
Crashes happen if we try to get PointerValues for unsupported types, and if the decls are not of a supported type, we can't collect anything useful.

Also skip compound assignments, which mostly aren't used with pointers and are better addressed for nullability purposes by requiring that pointer arithmetic is done only on Nonnull pointers.

PiperOrigin-RevId: 580970370
Change-Id: I03ef284762328ab72c422c89264efb78adadf9f9