Emit round-trip type annotations for C++ types.
A Rust type which was originally a C++ type gets an annotation to the effect of "I was originally, in C++, foo::Bar!". This is, for the moment, ignored. In a later CL, we can then see this, and when generating C++ bindings for the type, instead go "wait, no, nevermind, use the original type!"

We did discuss different ways to do this, in the end the conclusion we reached (AIUI) is that all the options are bad. I picked the easiest one to implement, and the one I view as most likely to be what we want to pursue in the long term. Since this is not something intended for end users, we still reserve the ability to rework it later.

(Also, honestly, even if we _did_ hand it out to end users, the refactoring to translate it to something else should be relatively straightforward.)

---

A quick note on structure:

The attribute, as written, is `__crubit::annotate`. This is mostly an implementation convenience: we can scan for all attributes named `__crubit::annotate` and search for our attribute in particular. Since, once again, this should not be exposed to end-users, it's fine. We're just picking whatever is easier for now.

If end users end up using these annotations, we might hide it behind a proc-macro, like `#[crubit::cc_type("xyz")]`, which expands to `#[__crubit::annotate(cc_type="...")]`

PiperOrigin-RevId: 584442104
Change-Id: I8f7416e1afcc75d72130894f99c4ecba1d0e20fd