rs_bindings_from_cc: Add reproducers for a C++ target with invalid character (for Rust crate name).
With 3 different cases:

A. With no Crubit in the picture. Useful for separating the issue between the Rust build rule and Crubit bazel integration.

B. Crubit enabled for the cc_library (with valid character) + Crubit-disabled for the cc_library with invalid character. Useful for assessing how much the invalid character is going to impact Crubit (direct) clients.

C. Crubit enabled for the cc_library with invalid character. Useful for understanding how Crubit (in the future) maps the invalid character.

PiperOrigin-RevId: 585905841
Change-Id: I82569db40d34e5b56517093346d0e5076925e8d2