Make the `string_view` bindings sound.
This removes `'static` reference conversions -- instead of being like a `&'static [u8]`, this change makes `string_view` more like a `*const [u8]`.

The old `'static` code makes sense from the point of view that an unsound C++ type can become an unsound Rust type. A `string_view` really is like a `&'static [u8]`, in that it can be passed around freely and used without any `unsafe` markers in C++! However, Crubit generally falls back to lifetimeless raw pointers, not to static references, when lifetime information is absent. So this treatment of `string_view` is inconsistent and of course dangerous.

(When it comes to _aliasing_, things are especially dangerous, and we've since decided to de-emphasize references in favor of (as yet undesigned) aliasing references anyway.)

PiperOrigin-RevId: 586139561
Change-Id: I3b6819b32cd93e983d073caef37622669d704af8