Round-trip types which were originally C++.
This really should basically just work. Its divided into two halves, which is a bit subtle, but they work together.

First, we insert the original C++ library as a `CcBindingsFromRustInfo` entry for the crate -- that is, we mark the original C++ library itself as the bindings for its Rust bindings. This later becomes a `--bindings-from-dependency` flag), we will always include all `public_hdrs` from the original C++ target. So we do actually include the headers that define the symbols, before we use them. (This is `prereqs.includes.extend(includes.iter().cloned());` in `format_ty_for_cc`.)

This fixes the includes.

Then, when actually using these libraries from C++, and have the right includes now, we substitute in the original C++ type name, using the attributes we've already begun populating in a prior change.

PiperOrigin-RevId: 586792462
Change-Id: I0a3be4f4bbc7bff97f76c9f26be6b8d9f15a95fe