Don't panic when encountering an incomplete record as parameter or return type.
This allows Crubit to keep generate bindings for other APIs that don't involve this incomplete record.

The header that references this incomplete record not through pointers or references is malformed and needs to fixed, but that can take time for a large codebase, therefore Crubit should be more lenient in this scenario.

We should probably surface a useful error message to indicate the exact reason as incomplete.

PiperOrigin-RevId: 586947378
Change-Id: I9c0ad06435c3a1eeb51c27adb0fa534c840a1c23