rs_bindings_from_cc: Map cc_library targets named 'core' to 'core_<last_package_path_component>'.
So that `::core` in generated bindings correctly resolves to Rust's 'core' crate.

While adding more integration tests for escaping core, I realized that Crubit's code generation needs to be made aware of the renaming of cc_library targets named 'core'  AND mapping of invalid characters to underscore as well, so I added logic for this.

PiperOrigin-RevId: 587140649
Change-Id: I0a8a94cb40d2391bc2e7234cc07e23ad4c320620