[nullability] Introduce a flag for smart pointer support (turned off by default).
The intent is to insulate the rest of the code from the changes being made for
smart pointers. See also comments in the code.

The implementation uses a global variable, which is generally frowned upon, but
I've tried to make it so that this global variable behaves as similar as
possible to a compile-time constant.

PiperOrigin-RevId: 587629717
Change-Id: I8508fa4fd765555a4b26926d5093d8a6af22cdb6