[nullability] Rename smart_pointers.cc to smart_pointers_diagnostics.cc.
In future patches, I want to introduce a new smart_pointers.cc that uses the
nullability_test testing style and then migrate most of the existing tests to
that. I also plan to write most of the new tests in that file. The intent is
that smart_pointers_diagnostics.cc will merely test the diagnostics-producing
part of the code.

PiperOrigin-RevId: 587633148
Change-Id: I7c1d56a6e81e4b08d502783ba37013fb0680b21f