[nullability] Make nullability_test output diagnostics to stderr.
Previously, these would be written to stdout, but I can't see this output
anywhere in my testing environment. I had a compile error in a new test that I'm
writing, but this only showed up as an assert failure when the analysis choked
on a recovery node in the AST. Because I didn't see the compile errors, this was
a bit confusing until I figured it out.

I plan to see if I can prevent the analysis from being run at all if there is a
compile error.

PiperOrigin-RevId: 587640654
Change-Id: I2bbc49e4831febdf9289b30825e6b32f517d95a8