[nullability] Fix bug in `underlyingRawPointerType()`.
The existing implementation didn't do the right thing when the smart pointer
type was wrapped in an alias template (which is very common, because it's how
our nullability annotations are implemented).

In this case, the line `T->getAs<TemplateSpecializationType>()` would find the
alias template, not the underlying smart pointer type that we wanted to find.

Adding a `.getCanonicalType()` after the `T` doesn't help either, because the
`TemplateSpecializationType` is sugar and therefore gets canonicalized away.

The solution is to, instead, check whether the `CXXRecordDecl` corresponding to
`T` is a `ClassTemplateSpecializationDecl` and retrieve the template arguments
that way.

This patch includes a test that fails without the fix.

PiperOrigin-RevId: 588143742
Change-Id: I54d4f8426944c0daabfa74809f29f5139d0fc116