Create an `extern_c` feature, which will get merged into `supported`.
We don't _need_ to deploy features this way -- we could just make `:supported` private and put things directly into it. However, I want to do a trial run for what it's like to do the proposed fine-grained feature release process, by deploying them gated behind a feature flag, and then flipping the flag in the BUILD file. Just, kinda, get a feel for what this process looks like.

The idea is:

1) add a bunch of things to `extern_c`
2) when they're all rolled out, and we have confidence, add `extern_c` to `_SUPPORTED_FEATURES`.
3) then, move the features to `supported`, and make `extern_c` empty.
4) finally, once that's rolled out, delete the no-op `extern_c` feature.

PiperOrigin-RevId: 588246155
Change-Id: I7e5edd099316f002a19de02a9a5972e01a07fb40