[nullability] Smart pointers: Make mock `unique_ptr` non-trivially-destructible.
This makes the shape of the AST look more like that of the real `unique_ptr`. In
particular, expressions containing temporaries now contain a
`CXXBindTemporaryExpr` node.

So far, we had not been handling `CXXBindTemporaryExpr` in the type transfer
function, and this causes tests to fail once `unique_ptr` becomes
non-trivially-destructible. This patch adds this handling to make tests pass
again.

PiperOrigin-RevId: 588692658
Change-Id: I8a089999865a1be77437257bfd98cc2688a963d3