[nullability] Smart pointers: Bug fix: Use `cast_or_null` instead of `cast`.
Because of the existence of `transferValue_SmartPointer()`, which creates a
location for any smart pointer glvalue that doesn't have one yet, I'm not
actually sure it's possible to hit this case. Still, I'd prefer the code to be
"locally nullability correct" (and not just because of the existence of this
global invariant).

PiperOrigin-RevId: 590089455
Change-Id: I94557674af1294320af59c2abd408436dd2e641a