[nullability] Smart pointers: Use `thisPointerType` matcher instead of `on`.
The `on` matcher would only match if the implicit object argument was a smart
pointer glvalue (i.e. if the method was called with the syntax `.method()`, but
not if it was a raw pointer to a smart pointer (i.e. if the method was called
with the syntax `->method()`).

This patch adds tests for the affected matchers that fail without the fix.
The patch also adds an "integration test" that uses the `->method()` syntax in a
test for the transfer function.

PiperOrigin-RevId: 590092576
Change-Id: I4def104e6bfdb9cf199fb5dfac2eea0c88e9bdd3