Move golden test logic to a bazel macro.
This has some noticeable simplifications to the golden/BUILD file, but the real reason I want to do it is to enable later writing golden tests that go in other directories -- in particular, to run these inside a top-level `examples/` directory. (By doing so, we can inject the contents of the goldens into documentation -- and also, this lets us cleanly separate out which goldens are for documentation/example purposes, and which are only for tests.)

This is not the only CL I need to do to make that work out-- I'd also want the `update.sh` script to magically work for goldens put somewhere else. And I probably want to add goldens for Rust, too, especially so that they can be used from docs. (And so, the golden_test macro might move into common.)

PiperOrigin-RevId: 592263844
Change-Id: Ifd4ab0a96e56804c3326b64e899eff6ad28ee288