Add functions taking/returning void/void* to the `extern "C"` subset.
This is, functionally, just enough to be able to have an integration test of it. That is, we add the notion of a "supported type" -- just void and pointers to supported types -- and a way to actually use them in bindings (functions).

From here on out, the increments can be smaller: we can add individual supported types, and separately add supported items (aliases, structs, unions, enums).

PiperOrigin-RevId: 592376616
Change-Id: I17862641fa96ca729dcbb0b40e44b05f4f5e2cf7