Suppress bindings for `nodiscard` and `deprecated` functions.
We should expect to map these to Rust `#[must_use]` and `#[deprecated]`, which should be easy enough after this CL -- most of this CL is just piping through _exactly_ the data you'd need in order to do that! But at the very last step, we stop short of that goal, and just exclude them from the `extern "C"` subset aimed at manually written bindings.

PiperOrigin-RevId: 592402956
Change-Id: I451c2f1d60d4f87782c81118f515aa8aa8350d67