rs_bindings_from_cc: Use a custom diagnoser when attempting to instantiate a class template specialization.
So that the diagnostics (for the fallable attempt of instantiation) doesn't gets reported as true failure, which causes Crubit to exit with error, which in turn causes the compilation to fail.

In short, Crubit attempts to instantiate class template specializations more eagerly than header parsing: e.g., `using A = MyTemplate<int>` in a header will cause Crubit to try to instantiate `MyTemplate<int>` while header parsing doesn't attempt to instantiate it. When `MyTemplate<int>` cannot be instantiated, Crubit will error out, while the cc_library target can still be built.

PiperOrigin-RevId: 594233384
Change-Id: Ia1d5fd799a85aebc55cf7c2150729afca4fd969e