Fail-closed on unknown C++ function attributes.
When an unknown C++ function attribute is encountered, we mark the function as `experimental`, since that attribute may affect how we want to expose bindings. This is a catch-all that excludes attributes like `enable_if`.

This may be sufficiently confusing that we should really actually report to the user what's going on. I've saved the attributes in a string, but a followup CL will actually output that string to the source file to explain the missing bindings. (I've also filed a bug and put a TODO to this effect -- since I haven't actually written that followup CL...)

PiperOrigin-RevId: 595242864
Change-Id: Iaee5a58121bb1cb12b25e9d4e63925ffc2423e4f