Fix for nullability widening bug.
This CL is a local fix for the general issue raised in https://github.com/llvm/llvm-project/issues/76462. It is an alternative to a pair of changes, one to the dataflow framework and a corresponding one here.

That pair of changes may still be a better long-term fix, but since this is simpler, arguably improves the code locally and doesn't require a framework fix, it seems preferable in the short term.

PiperOrigin-RevId: 595679336
Change-Id: I678044c0988e6db0d94db7b45a1b72281dda4c78