rs_bindings_from_cc: Use a custom diagnoser when attempting to deduce a class template specialization's method return type.
Similar to unknown commit, Crubit attempts to instantiate methods of class template specialization eagerly and attempts to deduce the method return type while doing so. This places stricter requirement on the class template specialization than 'being able to compile the C++ header', which causes Crubit to fail for targets that compile.

Note, while playing around, I noticed that this CL (expectedly) only helps with the failure in return type deduction, instead of the more general 'failing to instantiate a method' (and I have noted this in the comments).

PiperOrigin-RevId: 595782199
Change-Id: I3ab5aa67c084c938432242f4202f7c9ba8db8672