Support structs in `extern_c`.
In order to be both helpful to users and also easier to implement -- a rare double benefit -- we support any struct that has, in the abstract, supportable properties (e.g. is trivially relocatable etc.)  If there are unsupported fields, that's fine, they'll just be replaced with a blob.

Actually, this supports too many structs (e.g. structs with weird attributes, structs with nontrivially-destructible fields, templated structs, etc.), so will need to pare that down.

PiperOrigin-RevId: 596063221
Change-Id: I56856036e0bacb0a9c3ce2a01f62384785bfeb48