rs_bindings_from_cc: Move forward_declaration/ test into a subdirectory: included_before_definition/.
In doing so, I realized that the case where forward declarations and the definition in the same target to be un-interesting...so I have removed them. Additionally, since we have test case for handling multiple forward declarations from different targets, I've removed the test case for a single forward declaration in another target, to be more concise.

PiperOrigin-RevId: 597362381
Change-Id: I06287d4637e4c25b92fbbc65332a4684cfe9bb4f