Extremely rough cut refactoring: move `RsSnippet`, `RsTypeKind`, and so on to `rs_snippet.rs`.
There's a couple unrelated functions that needed to be moved as well, but I left that as a TODO to refine further.

The idea here is to, well, at least make things less enormous. I couldn't move all the tests over, but this is still something like 10% of the file, give or take. And even this CL was enough to _hang my IDE_ because apparently our tools don't scale to files 5-10KLOC long....

I think further refactorings are possible, in the realm of, like, moving generate_func to a different file than generate_record. That will be much more productive in terms of reducing the size of src_code_gen.rs. This attempt -- well, this CL makes me feel better about making RsTypeKind more detailed. But as an experiment for refactoring src_code_gen.rs, it was mostly a failure, sadly.

PiperOrigin-RevId: 597362421
Change-Id: I222b20626f2446343779b79b18166d36d07c8a5c