Simplify Cmdline by making it a struct, `CmdlineArgs`.
And satisfy the requirements of the style guide, around invariants, by wrapping it in a class, `Cmdline`.

(I think we discussed this a year or two ago and agreed a struct would be better, but the issue of how to make it style compliant was at stake. Well, I think this works well enough.)

I think the test `InstantiationsOutEmpty` (and the ones below it) is especially notable for how much more readable it's become. It was very unclear, before, what the test was doing. But because we now have a struct we mutate, I can separately initialize everything with valid values, and then specifically mutate it to be invalid, so now it's clear what the test is doing.

The main goal here was to try to make it remotely easy to add a new invocation, in particular for the crubit explorer endpoint. I haven't quite succeeded at that, but at least the most painful part is less painful now.

PiperOrigin-RevId: 599296132
Change-Id: I8d836f14d7f1f873e7c6c5545f165822b542635b