rs_bindings_from_cc: Canonicalize CXXRecord declarations by target.
Before this CL, when a forward declaration of `clang::CXXRecord` is first seen and thus is imported, we import its definition decl, if available. This causes the definition decl of a `clang::CXXRecord` to be attributed to the target of the forward declaration, which is not necessarily the same (and usually not the same) target as definition decl's target, which causes the bindings for the `clang::CXXRecord` to not be generated.

PiperOrigin-RevId: 599639528
Change-Id: Id474d8387708e6fa9d7f38ed86083e90585b1c0d