[nullability] Init nullability properties on smart pointers returned by reference.
Prior to this patch, this used to cause an assertion failure because we weren't
initializing the returned smart pointer correctly. The patch includes a test
in smart_pointers_diagnostics.cc that reproduces the assertion failure.

The patch also includes a test in smart_pointers.cc that fails (but does not
assert-fail) without the fix.

PiperOrigin-RevId: 600365759
Change-Id: Idf03f888427d5e74457fc93a4c1e66291a79ee92