Collect evidence from calls to CHECK(expr), a macro assumed to abort if !expr.
Adds necessary functionality to wrap the arguments to such macros in no-op functions that can be easily identified and noticed during evidence collection.

Uses this functionality in tests, but not yet in the `infer_tu_main` tool, which will need to explicitly engage the macro replacement functionality in order to collect any such evidence.

Also add already-passing test for verification handling of CHECK, i.e. treating the expression as Nonnull post-CHECK.

PiperOrigin-RevId: 601192616
Change-Id: Ided9f922132a66fa7bffd01f996981ef56317edb