Do not generate bindings for forward declarations of enums.
The output _before_ this CL is just incorrect -- it treats `enum X : int;` as equivalent to `enum X : int {};`. That is wrong, in that you can forward declare an enum that is defined in another target. We would be better off producing nothing for now.

The correct approach requires design work. Unlike other forward declared types, forward declared enums are complete types, and have a known size. This should be workable, or at least possible to make work, with `forward_declare.rs`, but at the moment we haven't done so.

PiperOrigin-RevId: 602615570
Change-Id: Icccadf42c3178bab2c8797694473af7c6936b51c